Refactor shared etl functions into separate module
Previously, code from `etl enrollment` was being imported into
`etl clinical` as a short-term solution to reduce code duplication.

Move these shared functions outside of both files into a shared module.
